web-1395: remove reloading & replace with prefetching links
